--- src/Components/SimpleElements.css
@@ -1,33 +1,0 @@
-.button_yellow {
-    background-color: rgb(255, 198, 23);
-    border-radius: 4px;
-    padding: 15px 25px;
-}
-
-.form_input {
-    border: none;
-    border-bottom: 1px solid rgb(234, 234, 234);
-    margin-bottom: 45px;
-}
-.form_input:focus {
-    border-bottom: 1px solid rgb(20, 115, 230);
-}
-
-.logo {
-    font-family: 'Forte';
-    font-size: 36px;
-    color: rgb(255, 198, 23);
-}
-.logo__img {
-    margin-right: 5px;
-}
-.logo__text_left {
-    margin-right: 5px;
-}
-.logo__text_right {
-    color: rgb(50, 50, 50);
-}
-
-.text_white {
-    color: white;
-}
