+++ src/Components/common/buttonActive/index.css
@@ -1,0 +1,5 @@
+.button_yellow {
+    background-color: rgb(255, 198, 23);
+    border-radius: 4px;
+    padding: 15px 25px;
+}
