+++ src/Components/common/buttonActive/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './index.css';
+
+const ButtonActive = ({ buttonClass, text }) => (
+    <div style={{ textAlign: 'right' }}>
+        <button className={`${buttonClass} button_yellow`}>{text}</button>
+    </div>
+);
+
+export default ButtonActive;
