+++ src/Components/common/inputText/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './index.css';
+
+const InputText = ({ inputClass, type, name, placeholder }) => (
+    <input
+        className={`${inputClass} form_input`}
+        type={type}
+        name={name}
+        placeholder={placeholder}
+    />
+);
+
+export default InputText;
