+++ src/Components/common/logo/index.css
@@ -1,0 +1,18 @@
+.logo {
+    font-family: 'Forte';
+    font-size: 36px;
+    color: rgb(255, 198, 23);
+}
+.logo__img {
+    margin-right: 5px;
+}
+.logo__text_left {
+    margin-right: 5px;
+}
+.logo__text_right {
+    color: rgb(50, 50, 50);
+}
+
+.text_white {
+    color: white;
+}
