+++ src/Components/layout/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../header';
+import Header from './header';
 
 import './index.css';
