+++ src/Components/profile/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 
 import './index.css';
 
-const Profile = ({ navigation }) => {
-    if (navigation !== 'profile') return null;
+const Profile = () => {
     return <h2>Профиль</h2>;
 };
